requirements.txt
openai
python-dotenv
notion-client
pytrends
snscrape
pandas
pyyaml
pytest
